refactor(drivers): use the new directory path in include
